requirements.txt
alembic==1.12.0
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.28.0
backoff==2.2.1
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.16.0
click==8.1.7
cryptography==41.0.4
dnspython==2.4.2
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.103.2
fastapi-filter==1.0.0
fastapi-pagination==0.12.10
fastapi-users==12.1.2
fastapi-users-db-sqlalchemy==6.0.1
greenlet==3.0.0
h11==0.14.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jose==1.0.0
makefun==1.15.1
Mako==1.2.4
MarkupSafe==2.1.3
orjson==3.9.7
passlib==1.7.4
pycparser==2.21
pydantic==2.4.2
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
PyJWT==2.8.0
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
six==1.16.0
sniffio==1.3.0
sqladmin==0.15.2
SQLAlchemy==2.0.22
starlette==0.27.0
typing_extensions==4.8.0
ujson==5.8.0
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
WTForms==3.0.1
